+++ s/index.node.ts
@@ -14,6 +14,6 @@ export * from "./transports/http/types.js"
 
 export * from "./transports/websocket/parts/connector.js"
 export * from "./transports/websocket/parts/integration.js"
-export * from "./transports/websocket/remote.js"
+export * from "./transports/websocket/client.js"
 export * from "./transports/websocket/types.js"