+++ s/transports/websocket/client.ts
@@ -3,27 +3,28 @@ import {once} from "@e280/stz"
 
 import {Fns} from "../../core/types.js"
 import {defaults} from "../defaults.js"
+import {WsRemoteOptions} from "./types.js"
 import {endpoint} from "../../core/endpoint.js"
-import {WebSocketRemoteOptions} from "./types.js"
 import {Messenger} from "../messenger/messenger.js"
 import {WebSocketConduit} from "../messenger/conduits/web-socket.js"
 import {waitForSocketOpen} from "../messenger/parts/wait-for-socket-open.js"
+import {Rtt} from "../../tools/pingponger.js"
 
-export async function webSocketRemote<ServerFns extends Fns>(
-		options: WebSocketRemoteOptions<ServerFns>
+export async function wsClient<ServerFns extends Fns>(
+		options: WsRemoteOptions<ServerFns>
 	) {
 
 	const {
 		tap,
 		socket,
 		rpc,
-		onDisconnect,
+		disconnected,
 		timeout = defaults.timeout,
 	} = options
 
 	await waitForSocketOpen(socket, timeout)
 
-	const disconnect = once(onDisconnect)
+	const disconnect = once(disconnected)
 
 	const conduit = new WebSocketConduit({
 		socket,
@@ -36,20 +37,31 @@ export async function webSocketRemote<ServerFns extends Fns>(
 		conduit,
 		tap,
 		timeout,
-		getLocalEndpoint: (serverside, rig) => endpoint({
+		getLocalEndpoint: (serverside) => endpoint({
 			tap,
-			fns: rpc(serverside, rig),
+			fns: rpc(serverside),
 		}),
 	})
 
-	const dispose = () => {
+	const detach = () => {
 		conduit.dispose()
 		messenger.dispose()
 	}
 
+	const close = () => {
+		detach()
+		socket.close()
+	}
+
+	// wait first ping result
+	await conduit.pingponger.onRtt.next()
+
 	return {
-		dispose,
+		close,
+		detach,
+		socket: options.socket,
 		remote: messenger.remote,
+		rtt: new Rtt(conduit.pingponger),
 	}
 }